+++ lib/webview/src/component/ChatInput.tsx
@@ -1,9 +1,13 @@
-import React, { useCallback } from "react";
+import React, { useCallback, useRef } from "react";
 
 export const ChatInput: React.FC<{
   placeholder?: string;
-  onSend: (message: string) => void;
-}> = ({ placeholder, onSend }) => {
+  disabled?: boolean;
+  text: string;
+  onChange?: (text: string) => void;
+  onEnter?: (text: string) => void;
+  onShiftEnter?: (text: string) => void;
+}> = ({ placeholder, disabled, text, onChange, onEnter, onShiftEnter }) => {
   // callback to automatically focus the input box
   const callbackRef = useCallback((inputElement: HTMLTextAreaElement) => {
     if (inputElement) {
@@ -14,17 +18,31 @@ export const ChatInput: React.FC<{
     }
   }, []);
 
+  const textareaWrapperRef = useRef<HTMLDivElement>(null);
+
   return (
-    <div className="chat-input">
+    <div className="chat-input" ref={textareaWrapperRef}>
       <textarea
+        disabled={disabled}
         ref={callbackRef}
         placeholder={placeholder}
-        rows={1 /* TODO auto-expand */}
-        onKeyUp={(event) => {
-          // ignore shift+enter to allow the user to enter multiple lines
+        value={text}
+        rows={1}
+        onInput={(event) => {
+          if (!textareaWrapperRef.current) return;
+
+          // `replicatedValue` is used in the CSS to expand the input
+          textareaWrapperRef.current.dataset.replicatedValue =
+            event.currentTarget.value;
+          onChange?.(event.currentTarget.value);
+        }}
+        // capture onKeyDown to prevent the user from adding enter to the input
+        onKeyDown={(event) => {
+          // enter sends enter, shift+enter creates a new line
           if (
+            onEnter != null &&
             event.key === "Enter" &&
-            !event.shiftKey &&
+            !(event.shiftKey || event.ctrlKey) &&
             event.target instanceof HTMLTextAreaElement
           ) {
             const value = event.target.value.trim();
@@ -32,7 +50,23 @@ export const ChatInput: React.FC<{
             if (value !== "") {
               event.preventDefault();
               event.stopPropagation();
-              onSend(value);
+              onEnter?.(value);
+            }
+          }
+
+          // shift-enter sends enter, enter creates a new line
+          if (
+            onShiftEnter != null &&
+            event.key === "Enter" &&
+            (event.shiftKey || event.ctrlKey) &&
+            event.target instanceof HTMLTextAreaElement
+          ) {
+            const value = event.target.value.trim();
+
+            if (value !== "") {
+              event.preventDefault();
+              event.stopPropagation();
+              onShiftEnter?.(value);
             }
           }
         }}